+++ models/WorkflowDefinition.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type WorkflowDefinition struct {
-	Id          uint64    `gorm:"column:id;primaryKey" json:"id"`
+	Id          int       `gorm:"column:id;primaryKey" json:"id"`
 	Title       string    `gorm:"column:title" json:"title"`
 	Description string    `gorm:"column:description" json:"description"`
 	CreateTime  time.Time `gorm:"column:create_time" json:"create_time"`
