+++ models/WorkflowInstance.go
@@ -3,10 +3,10 @@ package models
 import "time"
 
 type WorkflowInstance struct {
-	Id                   uint64    `gorm:"column:id" json:"id"`
-	WorkflowDefinitionId uint64    `gorm:"column:workflow_definition_id" json:"workflow_definition_id"`
-	CurrentStepId        uint64    `gorm:"column:current_step_id" json:"current_step_id"`
-	BusinessId           uint64    `gorm:"column:business_id" json:"business_id"`
+	Id                   int       `gorm:"column:id" json:"id"`
+	WorkflowDefinitionId int       `gorm:"column:workflow_definition_id" json:"workflow_definition_id"`
+	CurrentStepId        int       `gorm:"column:current_step_id" json:"current_step_id"`
+	BusinessId           int       `gorm:"column:business_id" json:"business_id"`
 	Status               string    `gorm:"column:status" json:"status"`
 	CreationTime         time.Time `gorm:"column:creation_time" json:"creation_time"`
 	Assignee             string    `gorm:"column:assignee" json:"assignee"`
